+++ node/js/shadow.js
@@ -33,7 +33,7 @@ export function useShadowValue(mirrorValue, { save, debounceMs, key } = {}) {
 
     useEffect(() => {
         if (!dirty.current) setValue(mirrorValue);
-        // eslint-disable-next-line react-hooks/exhaustive-deps
+         
     }, [mirrorValue]);
 
     const flush = async () => {
@@ -64,7 +64,7 @@ export function useShadowValue(mirrorValue, { save, debounceMs, key } = {}) {
             if (timer.current) clearTimeout(timer.current);
             if (dirty.current) flush();
         },
-        // eslint-disable-next-line react-hooks/exhaustive-deps
+         
         [key]
     );